[refactor] 모듈 Import 경로를 절대경로로 사용하도록 환경설정 및 디렉토리 리팩토링 - 모듈 import시 항상 절대경로로 import 하도록 환경설정 - 일부 모듈 폴더 구조 리팩토링
