+++ src/bitCoinChart/hooks/BinanceHooks.tsx
@@ -1,7 +1,7 @@
+import { CurrentPriceData } from '@bitCoinChart/types/CoinTypes';
 import { useQuery } from '@tanstack/react-query';
-import { CurrentPriceData } from '../types/CoinTypes';
 
-const useSymbolData = (symbol: string) => {
+export const useBinanceSymbolData = (symbol: string) => {
   return useQuery<CurrentPriceData | null>({
     queryKey: ['symbol', symbol],
     queryFn: async () => null,
@@ -9,5 +9,3 @@ const useSymbolData = (symbol: string) => {
     staleTime: Infinity, // worker를 통해 수동으로 적재하므로 staleTime을 높게 설정
   });
 };
-
-export default useSymbolData;
