+++ src/bitCoinChart/worker/binance/type/BinanceWorkerTypes.ts
