+++ src/bitCoinChart/worker/type/CoinCommonTypes.ts
