+++ src/bitCoinChart/worker/upbit/UpbitWorker.ts
@@ -1,10 +1,18 @@
+import { UPBIT_WEBSOCKET_URL } from '@bitCoinChart/types/CoinTypes';
+import { PriceMap } from '@bitCoinChart/worker/type/CoinCommonTypes';
+import { WorkerMessageEnum } from '@bitCoinChart/worker/enum/WorkerMessageEnum';
+import { UpbitTickerData } from '@bitCoinChart/worker/upbit/type/UpbitWorkerTypes';
+import {
+  fetchUpbitAllOpenPrices,
+  getUpbitAllSymbols,
+} from '@bitCoinChart/worker/upbit/util/UpbitWorkerUtils';
+import {
+  dataSetting,
+  fetchAllTickers,
+  getPriceColor,
+  isUsCountry,
+} from '@bitCoinChart/worker/util/WorkerUtils';
 import { v4 as uuidv4 } from 'uuid';
-import { PriceMap } from '../CoinCommonTypes';
-import { UpbitTickerData } from './UpbitWorkerTypes';
-import { fetchUpbitAllOpenPrices, getUpbitAllSymbols } from './UpbitWorkerUtils';
-import { dataSetting, fetchAllTickers, isUsCountry, getPriceColor } from '../WorkerUtils';
-import { UPBIT_WEBSOCKET_URL } from '../../types/CoinTypes';
-import { WorkerMessageEnum } from '../enum/WorkerMessageEnum';
 
 const priceMap: PriceMap = {};
 let ws: WebSocket | null = null;
