+++ src/bitCoinChart/worker/upbit/type/UpbitWorkerTypes.ts
