+++ src/components/ProjectNavigator.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import style from '../style/ProjectNavigator.module.scss';
+import style from '@/style/ProjectNavigator.module.scss';
 
 /**
  * 버튼을 통해 각 프로젝트로 진입할 수 있는 컴포넌트
