+++ src/components/ProjectRouter.tsx
@@ -1,9 +1,9 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import ProjectNavigator from './ProjectNavigator';
-import Header from './Header';
 import { isMobile } from 'react-device-detect';
 import { lazy, Suspense } from 'react';
-import LoadingFallback from './LoadingFallback';
+import LoadingFallback from '@components/LoadingFallback';
+import ProjectNavigator from '@components/ProjectNavigator';
+import Header from '@components/Header';
 
 const isGitHub = import.meta.env.VITE_DEPLOY_TARGET === 'GH';
 
@@ -13,7 +13,7 @@ const StopWatch = lazy(() => import('../stopwatch/components/Stopwatch'));
 // chart는 대부분 웹소켓 웹워커 쉐어드워커 등 무거운 작업을 하므로 스플리팅
 const CoinAppWrapper = lazy(() => import('../bitCoinChart/components/CoinAppWrapper'));
 const CoinApp = lazy(() => import('../bitCoinChart/components/CoinApp'));
-const CoinChartView = lazy(() => import('../bitCoinChart/components/CoinChartView'));
+const CoinChartView = lazy(() => import('../bitCoinChart/components/chart/CoinChartView'));
 /**
  * 라우터
  */
